pyproject.toml
[tool.poetry]
name = "cis188finalproj"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
selenium = "^4.1.3"
certifi = "^2021.10.8"
python-dotenv = "^0.20.0"
webdriver-manager = "^3.5.4"
pymongo = {version = "^4.1.1", extras = ["srv"]}
chromedriver-binary = "^101.0.4951"
beautifulsoup4 = "^4.11.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
